Backend/auth_service/requirements.txt
asyncpg
python-dotenv
fastapi
psycopg2